Separation of interface and implementation of analytics manager
